Create Find
This is a file that contains ways to traverse a div container in html.Inside the main,we want to  first highlight only those with class =item non-solid. So that is the first one. In the second main,we highlight those who are solid and in the third we highlight all the hot class.Same can be done for the cold.