Create Toggle visibility
With the toggle visibi;ity, I want to make sure that anytime the user clicks more info, a text appears. If he or she clicks more info again,the text disappears. So in this case we use the toggle in the main.js. It acts as a switch that turns it on and off